refactor(claw): move multipart's `File` definition from `ohkami_lib/` to `ohkami/`
